Added multiple functions to connection.php that make calls to stored procedures in the MySQL database. Functionality includes creating directories to store user photos, retrieving friend user information, creating an profile image relation tuple and establishing a reference to it in the user relation tuple.
Added a temporary home page. It now has a sign out button that redirects
the user to the login page and deletes any stored session data. Queried
SQL statements return information about the user's friends and display
the data to the webpage. This data shall be handled by JavaScript to
format it properly and give it some actual functionality.

Modified login and signup forms.