Added a new file 'getcontent.php' that handles sharing php session variables with javascript functions. The file home.php has been reworked to show the webpage when the proper user credentials are present in session variables. Otherwise, a message plus a blank page is rendered.
Chat thread messages are now retrieved from the database when a user
logs in. They are formatted with other information, including the
poster's profile image, full name and user name, text of the actual
message as well as date and time that message was recorded in database.

This branch is complete and these changes will be merged to the
master-branch. Another separate branch will be made in which message
submission will be implemented. This will likely involve cleaning user
input, using AJAX to send the input message to the server, creating a
new database record for the message with the necessary data. The data
will then be returned to the user and formatted using HTML and CSS. The
only issue is to ensure that any other active users have their webpage
updated to demonstrate these changes.