Created several React components, namely, the Home page component, which will stage all other smaller aspects of the web app like friends list, post feed, settings, chat threads, etc.
On the backend, set up Express to serve images as static files for easy
retrieval for profile pictures and such. Added some functionality that
allows the server to communicate with the MySQL server and retrieve data
and return it to the front end. Will add further functionality with
this, as chat messages between users need to be retrieved.

-Brian M.